requirements.txt
charset-normalizer==3.4.6
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
Faker==40.11.0
fonttools==4.62.1
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
reportlab==4.4.10
setuptools==82.0.1
six==1.17.0
ttkbootstrap==1.20.2
tzdata==2025.3